requirements.txt
ultralytics
rtmlib
opencv-python
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.7.0+cu118
torchvision==0.22.0+cu118
torchaudio==2.7.0+cu118
onnxruntime-gpu
tqdm
matplotlib
scipy
Pillow
numpy
prettytable
easydict
scikit-learn
pyyaml
yacs
termcolor
tabulate
tensorboard
gdown
faiss-gpu
faiss
lap
filterpy
SQLAlchemy
